Team_Workspace/Bae_JH/d_0318/Main_Docker_Runtime/requirements.txt
fastapi
uvicorn
pydantic

langchain
langchain-core
langchain-openai
openai
python-dotenv